[#96] Refactor: separate DeletAccountModal into file
